// src/types/fornecedores/FornecedorSupri.ts
export interface DadosFornecedor {
  razaoSocial: string;
  nomeFantasia: string;
  cnpj: string;
  inscricaoEstadual: string;
  email: string;
  telefone: string;
  celular: string;
  responsavel: string;
  site: string;
  tipoFornecedor: string;
  categoria: string;
  endereco: string;
  bairro: string;
  cidade: string;
  estado: string;
  cep: string;
  pais: string;
  observacoes: string;
  ativo: boolean;
  cliente_id: number;
}